Back at it again
This is a BUNCH of stuff, because i kind of let this project loose and forgot that the repository even existed, i got distracted and busy with Senai and a bit lazy to get back to this since i barely got the basics working.

hopefully i get back on track to clean this up. That would actually be nice