Refind the chat
For this commit i changed a bit about the NETWORK, MESSAGES and CHATBOX

New features:
* Login from cached credentials
* Cache credentials
* join a chat and talk to friends

To get that to work:
Instance 1
/login (or your login)
/host
*start talking

instance 2
/login (or your login)
/host
*start talking*

Next goal:
I want to get some simple UI going now that these simple sysyems are in place. Still needs changes tho, its not as stable as i would probably like it to be.