Some changes towards usability and stability WIP still
Main focus of this commit was at first, just to work on the Network.gd and make it as robust as possible. Altough it is not what happened as seen in this commit.

The Network code is not READY yet, i just copied and pasted the code from the docs, to make it easier to implement the code later on.

Now account caching is much better, if you login to another account it will automatically cache that account and you can refer back to login to it using the username...altough, if you didn't specify a username its a bit tricky yet, because there is no UI to actually tell what was the randomly generated username by nakama. Good enought for now, tho

The next step is just to make sure the code does not break easily with some tests, mainly Network wise.

Implement some simple UI components to display the accounts that are in Cache.

Implement some simple UI to display and access the Network API i made.

Not the best piece of code yet, but i'm pretty proud with the progress, because i know how things work for once in my life. I know its a bit messy, i might change it later, but so far i'm confident i can work this through.

Thank you God.